feat: add scalar util function
